only admins flush caches

+++ admin/actions.admin.php
@@ -328,9 +328,10 @@ new Store_RSS;
 
 add_action( 'admin_init', 'pagelines_set_versions' );
 function pagelines_set_versions() {
-	
-	delete_transient( 'pagelines_sections_cache' );
-	PageLinesRenderCSS::flush_version( false );
+	if ( current_user_can( 'edit_themes' ) ) {
+		delete_transient( 'pagelines_sections_cache' );
+		PageLinesRenderCSS::flush_version( false );
+	}
 	set_theme_mod( 'pagelines_version', pl_get_theme_data( get_template_directory(), 'Version' ) );
 	set_theme_mod( 'pagelines_child_version', pl_get_theme_data( get_stylesheet_directory(), 'Version' ) );
 }
